chore: :local 리패키징
