--- local/openmajor/src/main/java/com/suwiki/openmajor/converter/OpenMajor.kt
@@ -1,13 +1,0 @@
-package com.suwiki.openmajor.converter
-
-import com.suwiki.core.database.model.OpenMajorEntity
-import com.suwiki.core.model.OpenMajor
-
-fun OpenMajorEntity.toModel() = com.suwiki.core.model.OpenMajor(
-    id = id,
-    name = name,
-)
-
-fun com.suwiki.core.model.OpenMajor.toEntity() = OpenMajorEntity(name).also {
-    it.id = id
-}
