--- local/openmajor/src/main/java/com/suwiki/openmajor/datasource/LocalOpenMajorStorageDataSourceImpl.kt
@@ -1,19 +1,0 @@
-package com.suwiki.openmajor.datasource
-
-import com.suwiki.data.datasource.local.LocalOpenMajorStorageDataSource
-import com.suwiki.core.database.OpenMajorDatabase
-import com.suwiki.core.model.OpenMajor
-import com.suwiki.openmajor.converter.toEntity
-import javax.inject.Inject
-
-class LocalOpenMajorStorageDataSourceImpl @Inject constructor(
-    private val db: OpenMajorDatabase,
-) : LocalOpenMajorStorageDataSource {
-    override suspend fun saveAllOpenMajors(majors: List<com.suwiki.core.model.OpenMajor>) {
-        db.openMajorDao().insertAll(majors.map { it.toEntity() })
-    }
-
-    override suspend fun deleteAllOpenMajors() {
-        db.openMajorDao().deleteAll()
-    }
-}
