--- local/timetable/editor/src/main/java/com/suwiki/timetable/editor/converter/TimetableData.kt
@@ -1,19 +1,0 @@
-package com.suwiki.timetable.editor.converter
-
-import com.suwiki.core.database.model.TimetableEntity
-import com.suwiki.core.model.TimetableData
-
-fun TimetableData.toEntity(): TimetableEntity {
-    return TimetableEntity(
-        number,
-        major,
-        grade,
-        classNumber,
-        classDivideNumber,
-        className,
-        classification,
-        professor,
-        time,
-        credit,
-    )
-}
