+++ local/timetable/viewer/src/main/java/com/suwiki/local/timetable/viewer/converter/Timetable.kt
@@ -1,0 +1,17 @@
+package com.suwiki.local.timetable.viewer.converter
+
+import com.suwiki.core.database.model.TimetableEntity
+import com.suwiki.core.model.TimetableData
+
+fun TimetableEntity.toModel() = TimetableData(
+    number,
+    major,
+    grade,
+    classNumber,
+    classDivideNumber,
+    className,
+    classification,
+    professor,
+    time,
+    credit,
+)
