+++ local/timetable/viewer/src/main/java/com/suwiki/local/timetable/viewer/di/LocalDataSourceModule.kt
@@ -1,0 +1,20 @@
+package com.suwiki.local.timetable.viewer.di
+
+import com.suwiki.data.datasource.local.LocalTimetableProviderDatasource
+import com.suwiki.local.timetable.viewer.datasource.LocalTimetableProviderDatasourceImpl
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+abstract class LocalDataSourceModule {
+
+    @Singleton
+    @Binds
+    abstract fun bindLocalTimetableProviderDataSource(
+        localTimetableProviderDataSourceImpl: LocalTimetableProviderDatasourceImpl,
+    ): LocalTimetableProviderDatasource
+}
