--- local/token/storage/src/main/java/com/suwiki/token/storage/di/LocalDataSourceModule.kt
@@ -1,20 +1,0 @@
-package com.suwiki.token.storage.di
-
-import com.suwiki.data.datasource.local.LocalTokenStorageDataSource
-import com.suwiki.token.storage.datasource.LocalTokenStorageDataSourceImpl
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-abstract class LocalDataSourceModule {
-
-    @Singleton
-    @Binds
-    abstract fun bindLocalTokenStorageDataSource(
-        localTokenStorageDataSourceImpl: LocalTokenStorageDataSourceImpl,
-    ): LocalTokenStorageDataSource
-}
